backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cohere>=5.20.0",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "trafilatura>=2.0.0",
]
